user db에 프로필 사진 url 저장
