+++ src/pages/ProfilePhotoPage.jsx
@@ -1,24 +1,23 @@
 import React, {useState} from 'react';
 import {useForm} from 'react-hook-form';
 import ProfileUploader from '../components/ProfileUploader';
-import {useAuth} from '../hooks/useAuth';
 import Button from '../components/common/Button';
 import {compress} from '../utils/compress';
 import {useStorage} from '../hooks/useStorage';
 import {getURL} from '../firebase/storage';
+import {useUser} from '../hooks/useUser';
 
 export default function ProfilePhotoPage() {
   const {register, handleSubmit} = useForm();
-  const {user} = useAuth();
+  const {user, updatePhotoUrl} = useUser();
   const {upload} = useStorage(user?.uid);
   const [url, setPhotoUrl] = useState();
 
   const onSubmit = async ({profile}) => {
-    const [file] = profile;
     // 유저에 프로필 이미지를 설정한다는 것
     // 1. storage에 이미지를 저장
     // 2. 유저 db(firestore)에 이미지 주소를 저장
-
+    const [file] = profile;
     let photoUrl = '';
     // file이 있는 경우
     if (file) {
@@ -29,8 +28,9 @@ export default function ProfilePhotoPage() {
       const n = Math.ceil(Math.random() * 7);
       photoUrl = await getURL('default_image', n+'.png');
     }
-    setPhotoUrl(photoUrl);
     // 유저 db에 프로필 url 업데이트
+    setPhotoUrl(photoUrl);
+    updatePhotoUrl(photoUrl);
   };
 
   return (
